refactor(ui): move clear queue hook to separate file
